feat(team): 🛡️ disable role update for current user and prevent self-deletion
Disabled radio buttons for currently assigned roles in the change-role modal.
Prevented display of action menu for the current user by comparing profile ID.
