+++ src/queries/account.ts
@@ -21,6 +21,7 @@ export interface UserRole {
 }
 
 export interface User {
+  id: string
   email: string
   firstName: string
   lastName: string
